+++ lib/retriever.js
@@ -168,9 +168,11 @@ function retrieve(options, callback){
 
 
   function handleStreamError(record, err){
-    if(this.unpipe) this.unpipe();
-    if(this.destroy) this.destroy();
-    if(this.kill) this.kill();
+    if(this){
+      if(this.unpipe) this.unpipe();
+      if(this.destroy) this.destroy();
+      if(this.kill) this.kill();
+    }
     if(record._retrieverOutput){
       if(options.backupBucket){
         record._retrieverOutput.abortUpload(function(err){
@@ -186,10 +188,10 @@ function retrieve(options, callback){
 //split off backup, return stream for loader (or call loader)
   function handleStream(record, file, stream){
 
-    var pipeline = retrieverPipe(record, file, stream);
+    var pipeline = retrieverPipeline(record, file, stream);
 
     if(backupExists){
-      backup(options, stream, record, function(err){
+      backup(options, pipeline, record, function(err){
 
         backupsAttempted++;
 
@@ -206,8 +208,8 @@ function retrieve(options, callback){
       });
     }
 
-    loader(options, stream, record, function(err){
-      if(err) return handleStreamError.call(stream, record, err);
+    loader(options, pipeline, record, function(err){
+      if(err) return handleStreamError.call(pipeline, record, err);
       output.loaded.push(record.name);
 
       recordCallback(null, record);
